T16713 gc iterable

+++ phalcon/Session/Adapter/Stream.zep
@@ -95,16 +95,29 @@ class Stream extends Noop
      */
     public function gc(int max_lifetime) -> int|false
     {
-        var file, pattern, time;
+        var file, glob, last, pattern, time;
 
         let pattern = this->path . this->prefix . "*",
-            time    = time() - max_lifetime;
+            time    = time() - max_lifetime,
+            glob    = this->getGlobFiles(pattern);
 
-        for file in glob(pattern) {
-            if true === file_exists(file) &&
-               true === is_file(file)     &&
-               (filemtime(file) < time) {
-                unlink(file);
+        if (false === glob) {
+            let last = error_get_last();
+            if (isset(last["message"])) {
+                let last = last["message"];
+            } else {
+                let last = "Unexpected gc error";
+            }
+            throw new Exception(last);
+        }
+
+        if (!empty(glob)) {
+            for file in glob {
+                if true === file_exists(file) &&
+                   true === is_file(file)     &&
+                   (filemtime(file) < time) {
+                    unlink(file);
+                }
             }
         }
 
@@ -182,6 +195,26 @@ class Stream extends Noop
     private function getDirSeparator(string! directory) -> string
     {
         return rtrim(directory, DIRECTORY_SEPARATOR) . DIRECTORY_SEPARATOR;
+    }
+
+
+    /**
+     * Gets the glob array or returns false on failure
+     *
+     * @param string $pattern
+     *
+     * @return array|false
+     */
+    protected function getGlobFiles(string pattern) -> array | false
+    {
+        var errorLevel, glob;
+
+        let errorLevel = error_reporting(0);
+        error_clear_last();
+        let glob = glob(pattern);
+        error_reporting(errorLevel);
+
+        return glob;
     }
 
     /**
@@ -268,4 +301,5 @@ class Stream extends Noop
     {
         return is_writable(filename);
     }
+
 }
